+++ cpp/server/foobar2000/CMakeLists.txt
@@ -11,6 +11,7 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /d2notypeopt")
 set(
     FOOBAR2000_PLUGIN_SOURCES
     common.hpp
+    main_prefs_page.cpp main_prefs_page.hpp
     player.hpp
     player_control.cpp
     player_misc.cpp
@@ -21,7 +22,6 @@ set(
     prefs_page.cpp prefs_page.hpp
     resource.h resource.rc
     settings.cpp settings.hpp
-    settings_page.cpp settings_page.hpp
     utils.cpp utils.hpp
 )
