+++ cpp/server/foobar2000/main_prefs_page.cpp
@@ -1,4 +1,4 @@
-#include "settings_page.hpp"
+#include "main_prefs_page.hpp"
 #include "settings.hpp"
 #include "plugin.hpp"
 #include "resource.h"
