+++ cpp/server/foobar2000/main_prefs_page.hpp
